* new wjelement-VC100 project files to build using Visual Studio 2010 (I use Express Edition) * define macro WJE_DISTINGUISH_INTEGER_TYPE inside wjelement++.h * add 64bit support to the Visual Studio 2008 wjelement.vcproj project file (I use EE with Windows SDK 7.1A)

+++ src/wjelement/element.c
@@ -501,7 +501,7 @@ EXPORT XplBool _WJEWriteDocument(WJElement document, WJWriter writer, char *name
 					if (!current->value.number.negative) {
 						WJWUInt64(name, current->value.number.i, writer);
 					} else {
-						WJWInt64(name, -current->value.number.i, writer);
+						WJWInt64(name, -((int64) current->value.number.i), writer);
 					}
 				}
 				break;
